feat(eur-tauri): separate the launcher opening in separate code and add new function to window impl
