+++ crates/app/eur-tauri/src/main.rs
@@ -23,6 +23,9 @@ use eur_prompt_kit::PromptKitService;
 use eur_secret::secret;
 use eur_tauri::{
     WindowState, create_hover, create_launcher, create_window,
+    launcher::{
+        monitor_cursor_for_hover, open_launcher_window, position_hover_window, set_launcher_visible,
+    },
     procedures::{
         auth_procedures::{AuthApi, AuthApiImpl},
         chat_procedures::{ChatApi, ChatApiImpl},
@@ -45,102 +48,10 @@ use tauri::{
 };
 use tauri_plugin_global_shortcut::ShortcutState;
 use taurpc::Router;
-// Shared state to track if launcher is visible
-static LAUNCHER_VISIBLE: AtomicBool = AtomicBool::new(false);
-
 use tracing::{error, info};
 
 type SharedQuestionsClient = Arc<Mutex<Option<QuestionsClient>>>;
 type SharedPersonalDb = Arc<DatabaseManager>;
-
-/// Monitor information for window positioning
-#[derive(Debug, Clone)]
-struct MonitorInfo {
-    id: String,
-    x: i32,
-    y: i32,
-    width: u32,
-    height: u32,
-    scale_factor: f64,
-}
-
-/// Result of finding the monitor containing the cursor
-#[derive(Debug)]
-struct CursorMonitorResult {
-    monitor: MonitorInfo,
-    cursor_x: f64,
-    cursor_y: f64,
-}
-
-/// Find the monitor that contains the cursor position for WebviewWindow
-fn find_cursor_monitor(window: &tauri::WebviewWindow) -> Option<CursorMonitorResult> {
-    let cursor_position = window.cursor_position().ok()?;
-    find_cursor_monitor_impl(cursor_position)
-}
-
-/// Find the monitor that contains the cursor position for Window
-fn find_cursor_monitor_window(window: &tauri::Window) -> Option<CursorMonitorResult> {
-    let cursor_position = window.cursor_position().ok()?;
-    find_cursor_monitor_impl(cursor_position)
-}
-
-/// Internal implementation for finding cursor monitor
-fn find_cursor_monitor_impl(
-    cursor_position: tauri::PhysicalPosition<f64>,
-) -> Option<CursorMonitorResult> {
-    let monitors = get_all_monitors().ok()?;
-
-    for monitor in monitors {
-        let monitor_id = monitor.id().unwrap_or_default().to_string();
-        let scale_factor = monitor.scale_factor().unwrap_or(1.0) as f64;
-        let monitor_width = (monitor.width().unwrap_or(1920) as f64 * scale_factor) as u32;
-        let monitor_height = (monitor.height().unwrap_or(1080) as f64 * scale_factor) as u32;
-        let monitor_x = (monitor.x().unwrap_or(0) as f64 * scale_factor) as i32;
-        let monitor_y = (monitor.y().unwrap_or(0) as f64 * scale_factor) as i32;
-
-        // Check if cursor is on this monitor
-        if cursor_position.x >= monitor_x as f64
-            && cursor_position.x <= (monitor_x + monitor_width as i32) as f64
-            && cursor_position.y >= monitor_y as f64
-            && cursor_position.y <= (monitor_y + monitor_height as i32) as f64
-        {
-            return Some(CursorMonitorResult {
-                monitor: MonitorInfo {
-                    id: monitor_id,
-                    x: monitor_x,
-                    y: monitor_y,
-                    width: monitor_width,
-                    height: monitor_height,
-                    scale_factor,
-                },
-                cursor_x: cursor_position.x,
-                cursor_y: cursor_position.y,
-            });
-        }
-    }
-    None
-}
-
-/// Calculate launcher position (centered horizontally, 1/4 from top)
-fn calculate_launcher_position(
-    monitor: &MonitorInfo,
-    window_size: tauri::PhysicalSize<u32>,
-) -> (i32, i32) {
-    let launcher_x = monitor.x + (monitor.width as i32 - window_size.width as i32) / 2;
-    let launcher_y = monitor.y + (monitor.height as i32 - window_size.height as i32) / 4;
-    (launcher_x, launcher_y)
-}
-
-/// Calculate hover window position (right side, 3/4 down)
-fn calculate_hover_position(
-    monitor: &MonitorInfo,
-    window_size: tauri::PhysicalSize<u32>,
-) -> (i32, i32) {
-    let hover_x = monitor.x + monitor.width as i32 - window_size.width as i32 - 10; // 10px margin from edge
-    let hover_y =
-        monitor.y + (monitor.height as f64 * 0.75) as i32 - (window_size.height as i32 / 2);
-    (hover_x, hover_y)
-}
 
 async fn create_shared_database_manager(app_handle: &tauri::AppHandle) -> SharedPersonalDb {
     let db_path = get_db_path(app_handle);
@@ -396,7 +307,7 @@ fn main() {
                                     launcher
                                         .emit("launcher_closed", ())
                                         .expect("Failed to emit launcher_closed event");
-                                    LAUNCHER_VISIBLE.store(false, Ordering::SeqCst);
+                                    set_launcher_visible(false);
                                     // Ensure state is updated
                                 }
                             }
@@ -446,7 +357,7 @@ fn main() {
                                 window
                                     .emit("launcher_closed", ())
                                     .expect("Failed to emit launcher_closed event");
-                                LAUNCHER_VISIBLE.store(false, Ordering::SeqCst);
+                                set_launcher_visible(false);
                                 // Ensure state is updated
                             }
                         }
@@ -507,187 +418,15 @@ fn shortcut_plugin(launcher_label: String) -> TauriPlugin<Wry> {
                     .expect("Failed to emit launcher_closed event");
 
                 // Update the shared state to indicate launcher is hidden
-                LAUNCHER_VISIBLE.store(false, Ordering::SeqCst);
+                set_launcher_visible(false);
             } else {
-                // Update the shared state to indicate launcher is visible
-                LAUNCHER_VISIBLE.store(true, Ordering::SeqCst);
-
-                // Use consolidated monitor detection function
-                if let Some(cursor_monitor) = find_cursor_monitor_window(&launcher) {
-                    let monitor = &cursor_monitor.monitor;
-                    let window_size = launcher.inner_size().unwrap();
-
-                    info!("Monitor width: {:?}", monitor.width);
-                    info!("Monitor height: {:?}", monitor.height);
-                    info!("Monitor x: {:?}", monitor.x);
-                    info!("Monitor y: {:?}", monitor.y);
-                    info!("Monitor scale factor: {:?}", monitor.scale_factor);
-                    info!("Window size: {:?}", window_size);
-
-                    // Calculate launcher position using consolidated function
-                    let (launcher_x, launcher_y) =
-                        calculate_launcher_position(monitor, window_size);
-
-                    info!("Launcher position: ({}, {})", launcher_x, launcher_y);
-
-                    launcher
-                        .set_position(tauri::Position::Physical(tauri::PhysicalPosition {
-                            x: launcher_x,
-                            y: launcher_y,
-                        }))
-                        .expect("Failed to set launcher position");
-
-                    // Calculate relative position for screen capture
-                    let capture_x = ((monitor.width as i32 as f64) / 2.0) as i32
-                        - (window_size.width as f64 / 2.0) as i32;
-                    let capture_y = ((monitor.height as i32 as f64) / 4.0) as i32
-                        - (window_size.height as f64 / 4.0) as i32;
-                    let start_record = std::time::Instant::now();
-                    // Capture the screen region behind the launcher
-                    match capture_focused_region_rgba(
-                        monitor.id.clone(),
-                        capture_x as u32,
-                        capture_y as u32,
-                        window_size.width,
-                        window_size.height,
-                    ) {
-                        Ok(img) => {
-                            let t0 = std::time::Instant::now();
-                            let img = image::DynamicImage::ImageRgba8(img.clone()).to_rgb8();
-
-                            info!("Captured image size: {:?}", img.dimensions());
-                            let duration = t0.elapsed();
-                            info!("Capture of background area completed in: {:?}", duration);
-
-                            // Convert the image to base64
-                            if let Ok(base64_image) = image_to_base64(img) {
-                                // Send the base64 image to the frontend
-                                launcher
-                                    .emit("background_image", base64_image)
-                                    .expect("Failed to emit background_image event");
-                            }
-                        }
-                        Err(e) => {
-                            error!("Failed to capture screen region: {}", e);
-                        }
-                    }
-                    let duration = start_record.elapsed();
-                    info!("Capture of background area completed in: {:?}", duration);
-
-                    // Only show the launcher if it was previously hidden
-                    launcher.show().expect("Failed to show launcher window");
-
-                    // Emit an event to notify that the launcher has been opened
-                    // Include positioning information for proper background alignment
-                    let launcher_info = serde_json::json!({
-                        "monitor_id": monitor.id.clone(),
-                        "launcher_x": launcher_x,
-                        "launcher_y": launcher_y,
-                        "launcher_width": window_size.width,
-                        "launcher_height": window_size.height,
-                        "monitor_width": monitor.width,
-                        "monitor_height": monitor.height
-                    });
-                    launcher
-                        .emit("launcher_opened", launcher_info)
-                        .expect("Failed to emit launcher_opened event");
-
-                    launcher
-                        .set_focus()
-                        .expect("Failed to focus launcher window");
+                // Use the extracted launcher opening function
+                if let Err(e) = open_launcher_window(&launcher) {
+                    error!("Failed to open launcher window: {}", e);
                 }
             }
         })
         .build()
-}
-
-/// Position the hover window to the right side, around 3/4 to the bottom of the screen
-fn position_hover_window(hover_window: &tauri::WebviewWindow) {
-    if let Some(cursor_monitor) = find_cursor_monitor(hover_window) {
-        let monitor = &cursor_monitor.monitor;
-        let window_size = hover_window.inner_size().unwrap_or(tauri::PhysicalSize {
-            width: 50,
-            height: 50,
-        });
-
-        // Calculate hover position using consolidated function
-        let (hover_x, hover_y) = calculate_hover_position(monitor, window_size);
-
-        info!(
-            "Positioning hover window at: ({}, {}) on monitor {}x{}",
-            hover_x, hover_y, monitor.width, monitor.height
-        );
-
-        if let Err(e) =
-            hover_window.set_position(tauri::Position::Physical(tauri::PhysicalPosition {
-                x: hover_x,
-                y: hover_y,
-            }))
-        {
-            error!("Failed to set hover window position: {}", e);
-        }
-    }
-}
-
-/// Monitor cursor position and reposition hover window when cursor moves to different screen
-async fn monitor_cursor_for_hover(hover_window: tauri::WebviewWindow) {
-    let mut last_monitor_id = String::new();
-    let mut last_cursor_x = 0.0;
-    let mut last_cursor_y = 0.0;
-
-    loop {
-        // Very fast polling for maximum responsiveness - check every 16ms (~60fps)
-        tokio::time::sleep(tokio::time::Duration::from_millis(100)).await;
-
-        if let Ok(cursor_position) = hover_window.cursor_position() {
-            // Only proceed if cursor actually moved (avoid unnecessary work)
-            if (cursor_position.x - last_cursor_x).abs() < 1.0
-                && (cursor_position.y - last_cursor_y).abs() < 1.0
-            {
-                continue;
-            }
-
-            last_cursor_x = cursor_position.x;
-            last_cursor_y = cursor_position.y;
-
-            // Use consolidated monitor detection function
-            if let Some(cursor_monitor) = find_cursor_monitor(&hover_window) {
-                let monitor = &cursor_monitor.monitor;
-
-                // If cursor moved to a different monitor, reposition hover window immediately
-                if monitor.id != last_monitor_id {
-                    info!(
-                        "Cursor moved to monitor: {} (immediate repositioning)",
-                        monitor.id
-                    );
-                    last_monitor_id = monitor.id.clone();
-
-                    // Position hover window on the new monitor
-                    let window_size = hover_window.inner_size().unwrap_or(tauri::PhysicalSize {
-                        width: 50,
-                        height: 50,
-                    });
-
-                    // Calculate hover position using consolidated function
-                    let (hover_x, hover_y) = calculate_hover_position(monitor, window_size);
-
-                    info!(
-                        "Repositioning hover window to: ({}, {}) on monitor {}x{}",
-                        hover_x, hover_y, monitor.width, monitor.height
-                    );
-
-                    if let Err(e) = hover_window.set_position(tauri::Position::Physical(
-                        tauri::PhysicalPosition {
-                            x: hover_x,
-                            y: hover_y,
-                        },
-                    )) {
-                        error!("Failed to reposition hover window: {}", e);
-                    }
-                }
-            }
-        }
-    }
 }
 
 #[tauri::command]
